util,mock: move utils.MockCluster to mock.Cluster
