+++ pkg/restore/db_test.go
@@ -13,18 +13,19 @@ import (
 	"github.com/pingcap/tidb/util/testleak"
 
 	"github.com/pingcap/br/pkg/gluetidb"
+	"github.com/pingcap/br/pkg/mock"
 	"github.com/pingcap/br/pkg/utils"
 )
 
 var _ = Suite(&testRestoreSchemaSuite{})
 
 type testRestoreSchemaSuite struct {
-	mock *utils.MockCluster
+	mock *mock.Cluster
 }
 
 func (s *testRestoreSchemaSuite) SetUpSuite(c *C) {
 	var err error
-	s.mock, err = utils.NewMockCluster()
+	s.mock, err = mock.NewCluster()
 	c.Assert(err, IsNil)
 }
 func TestT(t *testing.T) {
